Restructure ListeningEvents feature tests
